new function and re-use tokens
added 2nd left button makes a special gps point push
adding a agoTOKEN.ini to hold token for ago. Was way too wasteful to be
requesting a new token every 20 seconds. Does a check on program startup
for valid token, if not, get a new one and put it in the file.